fixed not finding ROM in some situations

+++ Convert2Dualboot-SD-DOS/C2DBSD.cmd
@@ -9,7 +9,7 @@
 ::#Setup the menu for the batch file
 :menu
 cls
-echo.
+echo.&echo.&echo.&echo.&echo.&echo.&
 echo ********************* Convert2Dualboot-SD for DOS **********************
 echo.
 echo  A tool to modify standard flashable ROM zips for Racks11479 DualbootSD
@@ -21,7 +21,7 @@ echo   4    Prep Gapps for DualbootSD Alternate Boot
 echo   5    Clear out recent mods
 echo   0    Quit
 echo.
-echo **************************************************************************
+echo ********************* Convert2Dualboot-SD for DOS **********************
 echo.
 echo Please make your decision: 
 choice /C 123450 /N 
@@ -43,40 +43,47 @@ set PREPATH=%PATH%
 set PATH=%PATH%;.\tools;..\tools;..\..\tools
 mkdir pri-to-modify\gapps
 cls
-FOR /F %%R IN ('DIR/B/ON "pri-to-modify\*gapps*.zip"') DO mv -f pri-to-modify/%%R pri-to-modify/gapps/%%R
-FOR /F %%R IN ('DIR/B/ON "pri-to-modify\*.zip"') DO set ROM=%%R
-unzip pri-to-modify/%ROM% ramdisk.img -d tmp
-unzip pri-to-modify/%ROM% META-INF/com/google/android/updater-script -d tmp
-unzip pri-to-modify/%ROM% system/etc/vold.fstab -d tmp
-cd tmp\rd
-dd if=../ramdisk.img bs=64 skip=1 of=ramdisk
-gunzip -c ramdisk | cpio -i
-rm -r ramdisk
-sed -i s/mmcblk0p5/mmcblk1p2/ init.encore.rc 
-sed -i s/mmcblk0p6/mmcblk1p3/ init.encore.rc
-cd ..
-rm ramdisk.img
-mkbootfs rd | gzip -9 > nuRamdisk-new.gz
-echo.
-
-mkimage -A ARM -T RAMDisk -n Image -d nuRamdisk-new.gz ramdisk.img
-echo.
-
-rmdir /S /Q rd
-rm nuRamdisk-new.gz
-
-sed -i "s/mmcblk0p5/mmcblk1p2/" META-INF/com/google/android/updater-script
-sed -i "s/mmcblk0p1/mmcblk1p1/" META-INF/com/google/android/updater-script
-sed -i "s,/system,/system1,g" META-INF/com/google/android/updater-script  
-
-sed -i "s/sdcard auto/sdcard 7/" system/etc/vold.fstab
-cp -f ../pri-to-modify/%ROM% ../pri-to-modify/update_RDBSD_pri_%ROM%
-zip -R -u ../pri-to-modify/update_RDBSD_pri_%ROM% *.*
-cd ..
-mv -f pri-to-modify/update_RDBSD_pri_%ROM% Primary-Mod
-mv -f pri-to-modify/gapps/*.zip pri-to-modify
-rm -r tmp
-rm -r pri-to-modify/gapps
+if exist pri-to-modify\*gapp*.zip FOR /F %%R IN ('DIR/B/ON "pri-to-modify\*gapps*.zip"') DO mv -f pri-to-modify/%%R pri-to-modify/gapps/%%R >nul
+if exist pri-to-modify\*.zip FOR /F %%R IN ('DIR/B/ON pri-to-modify\*.zip') DO set ROM=%%R
+If NOT exist pri-to-modify\*.zip (
+	cls
+	echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.
+	echo ***** ROM file not found *****
+	sleep 10s
+	) Else (
+	unzip -oq pri-to-modify/%ROM% ramdisk.img -d tmp
+	unzip -oq pri-to-modify/%ROM% META-INF/com/google/android/updater-script -d tmp
+	unzip -oq pri-to-modify/%ROM% system/etc/vold.fstab -d tmp
+	cd tmp\rd
+	dd if=../ramdisk.img bs=64 skip=1 of=ramdisk > nul
+	gunzip -c ramdisk | cpio -i > nul
+	rm -r ramdisk > nul
+	sed -i s/mmcblk0p5/mmcblk1p2/ init.encore.rc
+	sed -i s/mmcblk0p6/mmcblk1p3/ init.encore.rc
+	cd ..
+	rm ramdisk.img > nul
+	mkbootfs rd | gzip -9 > nuRamdisk-new.gz
+	echo.
+	
+	mkimage -A ARM -T RAMDisk -n Image -d nuRamdisk-new.gz ramdisk.img > nul
+	echo.
+	
+	rmdir /S /Q rd
+	rm nuRamdisk-new.gz
+	
+	sed -i s/mmcblk0p5/mmcblk1p2/ META-INF/com/google/android/updater-script
+	sed -i s/mmcblk0p1/mmcblk1p1/ META-INF/com/google/android/updater-script
+	sed -i "s,/system,/system1,g" META-INF/com/google/android/updater-script
+	
+	sed -i "s/sdcard auto/sdcard 7/" system/etc/vold.fstab
+	cp -f ../pri-to-modify/%ROM% ../pri-to-modify/update_RDBSD_pri_%ROM% > nul
+	zip -ruq ../pri-to-modify/update_RDBSD_pri_%ROM% *.*
+	cd ..
+	mv -f pri-to-modify/update_RDBSD_pri_%ROM% Primary-Mod > nul
+	)
+if exist pri-to-modify\gapps\*.zip mv -f pri-to-modify/gapps/*.zip pri-to-modify > nul
+rm -fr tmp
+rm -fr pri-to-modify/gapps
 set PATH=%PREPATH%
 set PREPATH=
 echo.
@@ -91,42 +98,49 @@ set PREPATH=%PATH%
 set PATH=%PATH%;.\tools;..\tools;..\..\tools
 mkdir alt-to-modify\gapps
 cls
-FOR /F %%R IN ('DIR/B/ON "alt-to-modify\*gapps*.zip"') DO mv -f alt-to-modify/%%R alt-to-modify/gapps/%%R
-FOR /F %%R IN ('DIR/B/ON "alt-to-modify\*.zip"') DO set ROM=%%R
-unzip alt-to-modify/%ROM% ramdisk.img -d tmp
-unzip alt-to-modify/%ROM% META-INF/com/google/android/updater-script -d tmp
-unzip alt-to-modify/%ROM% system/etc/vold.fstab -d tmp
-cd tmp\rd
-dd if=../ramdisk.img bs=64 skip=1 of=ramdisk
-gunzip -c ramdisk | cpio -i
-rm -r ramdisk
-sed -i s/mmcblk0p5/mmcblk1p2/ init.encore.rc 
-sed -i s/mmcblk0p6/mmcblk1p3/ init.encore.rc
-cd ..
-rm ramdisk.img
-mkbootfs rd | gzip -9 > nuRamdisk-new.gz
-echo.
-
-mkimage -A ARM -T RAMDisk -n Image -d nuRamdisk-new.gz ramdisk.img
-echo.
-
-rmdir /S /Q rd
-rm nuRamdisk-new.gz
-
-sed -i "s/mmcblk0p5/mmcblk1p5/"  META-INF/com/google/android/updater-script
-sed -i "s/mmcblk0p1/mmcblk1p1/"  META-INF/com/google/android/updater-script
-sed -i "s,/system,/system2,g"  META-INF/com/google/android/updater-script
-sed -i "s/uImage/uAltImg/"  META-INF/com/google/android/updater-script
-sed -i "s/uRamdisk/uAltRam/"  META-INF/com/google/android/updater-script
-
-sed -i "s/sdcard auto/sdcard 7/" system/etc/vold.fstab
-cp -f ../alt-to-modify/%ROM% ../alt-to-modify/update_RDBSD_alt_%ROM%
-zip -R -u ../alt-to-modify/update_RDBSD_alt_%ROM% *.*
-cd ..
-mv -f alt-to-modify/update_RDBSD_alt_%ROM% Alternate-Mod
-mv -f alt-to-modify/gapps/*.zip alt-to-modify
-rm -r tmp
-rm -r alt-to-modify/gapps
+if exist alt-to-modify\*gapp*.zip FOR /F %%R IN ('DIR/B/ON "alt-to-modify\*gapps*.zip"') DO mv -f alt-to-modify/%%R alt-to-modify/gapps/%%R >nul
+if exist alt-to-modify\*.zip FOR /F %%R IN ('DIR/B/ON alt-to-modify\*.zip') DO set ROM=%%R
+If NOT exist alt-to-modify\*.zip (
+	cls
+	echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.
+	echo ***** ROM file not found *****
+	sleep 10s
+	) Else (
+	unzip -oq alt-to-modify/%ROM% ramdisk.img -d tmp
+	unzip -oq alt-to-modify/%ROM% META-INF/com/google/android/updater-script -d tmp
+	unzip -oq alt-to-modify/%ROM% system/etc/vold.fstab -d tmp
+	cd tmp\rd
+	dd if=../ramdisk.img bs=64 skip=1 of=ramdisk > nul
+	gunzip -c ramdisk | cpio -i > nul
+	rm -r ramdisk > nul
+	sed -i s/mmcblk0p5/mmcblk1p2/ init.encore.rc
+	sed -i s/mmcblk0p6/mmcblk1p3/ init.encore.rc
+	cd ..
+	rm ramdisk.img > nul
+	mkbootfs rd | gzip -9 > nuRamdisk-new.gz
+	echo.
+	
+	mkimage -A ARM -T RAMDisk -n Image -d nuRamdisk-new.gz ramdisk.img > nul
+	echo.
+	
+	rmdir /S /Q rd
+	rm nuRamdisk-new.gz
+	
+	sed -i s/mmcblk0p5/mmcblk1p5/ META-INF/com/google/android/updater-script
+	sed -i s/mmcblk0p1/mmcblk1p1/ META-INF/com/google/android/updater-script
+	sed -i "s,/system,/system2,g" META-INF/com/google/android/updater-script
+	sed -i s/uImage/uAltImg/ META-INF/com/google/android/updater-script
+	sed -i s/uRamdisk/uAltRam/ META-INF/com/google/android/updater-script
+	
+	sed -i "s/sdcard auto/sdcard 7/" system/etc/vold.fstab
+	cp -f ../alt-to-modify/%ROM% ../alt-to-modify/update_RDBSD_alt_%ROM% > nul
+	zip -ruq ../alt-to-modify/update_RDBSD_alt_%ROM% *.*
+	cd ..
+	mv -f alt-to-modify/update_RDBSD_alt_%ROM% Alternate-Mod > nul
+	)
+if exist pri-to-modify\gapps\*.zip mv -f pri-to-modify/gapps/*.zip pri-to-modify > nul
+rm -fr tmp
+rm -fr alt-to-modify/gapps
 set PATH=%PREPATH%
 set PREPATH=
 echo.
@@ -136,25 +150,30 @@ goto :menu
 ::#Modify Gapps for Primary Boot
 :gp
 mkdir tmp
-mkdir tmp\rd
 set PREPATH=%PATH%
 set PATH=%PATH%;.\tools;..\tools;..\..\tools
 cls
 FOR /F %%R IN ('DIR/B/ON "pri-to-modify\gapps*.zip"') DO set GAPPS=%%R
-unzip pri-to-modify/%GAPPS% META-INF/com/google/android/updater-script -d tmp
-unzip pri-to-modify/%GAPPS% install-optional.sh -d tmp
-
-cd tmp
-
-sed -i "s,/system,/system1,g" META-INF/com/google/android/updater-script
-sed -i "s,/system,/system1,g" install-optional.sh
-
-rmdir /S /Q rd
-cp -f ../pri-to-modify/%GAPPS% ../pri-to-modify/gapps_RDBSD_pri_%GAPPS%
-zip -R -u ../pri-to-modify/gapps_RDBSD_pri_%GAPPS% *.*
-mv -f ../pri-to-modify/gapps_RDBSD_pri_%GAPPS% ../Primary-Mod/gapps_RDBSD_pri_%GAPPS%
-cd ..
-rm -r tmp
+If NOT exist pri-to-modify\%GAPPS% (
+	cls
+	echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.
+	echo ***** GApps file not found *****
+	sleep 10s
+	) Else (
+	unzip -oq pri-to-modify/%GAPPS% META-INF/com/google/android/updater-script -d tmp
+	unzip -oq pri-to-modify/%GAPPS% install-optional.sh -d tmp
+	
+	cd tmp
+	
+	sed -i "s,/system,/system1,g" META-INF/com/google/android/updater-script
+	sed -i "s,/system,/system1,g" install-optional.sh
+	
+	cp -f ../pri-to-modify/%GAPPS% ../pri-to-modify/gapps_RDBSD_pri_%GAPPS%
+	zip -ruq ../pri-to-modify/gapps_RDBSD_pri_%GAPPS% *.*
+	mv -f ../pri-to-modify/gapps_RDBSD_pri_%GAPPS% ../Primary-Mod/gapps_RDBSD_pri_%GAPPS%
+	cd ..
+	)
+rm -fr tmp
 set PATH=%PREPATH%
 set PREPATH=
 echo.
@@ -164,25 +183,30 @@ goto :menu
 ::#Modify Gapps for Alternate Boot
 :ga
 mkdir tmp
-mkdir tmp\rd
 set PREPATH=%PATH%
 set PATH=%PATH%;.\tools;..\tools;..\..\tools
 cls
-FOR /F %%R IN ('DIR/B/ON "pri-to-modify\gapps*.zip"') DO set GAPPS=%%R
-unzip pri-to-modify/%GAPPS% META-INF/com/google/android/updater-script -d tmp
-unzip pri-to-modify/%GAPPS% install-optional.sh -d tmp
-
-cd tmp
-
-sed -i "s,/system,/system1,g" META-INF/com/google/android/updater-script
-sed -i "s,/system,/system1,g" install-optional.sh
-
-rmdir /S /Q rd
-cp -f ../pri-to-modify/%GAPPS% ../pri-to-modify/gapps_RDBSD_pri_%GAPPS%
-zip -R -u ../pri-to-modify/gapps_RDBSD_pri_%GAPPS% *.*
-mv -f ../pri-to-modify/gapps_RDBSD_pri_%GAPPS% ../Primary-Mod/gapps_RDBSD_pri_%GAPPS%
-cd ..
-rm -r tmp
+FOR /F %%R IN ('DIR/B/ON "alt-to-modify\gapps*.zip"') DO set GAPPS=%%R
+If NOT exist alt-to-modify\%GAPPS% (
+	cls
+	echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.
+	echo ***** GApps file not found *****
+	sleep 10s
+	) Else (
+	unzip oq alt-to-modify/%GAPPS% META-INF/com/google/android/updater-script -d tmp
+	unzip oq alt-to-modify/%GAPPS% install-optional.sh -d tmp
+	
+	cd tmp
+	
+	sed -i "s,/system,/system2,g" META-INF/com/google/android/updater-script
+	sed -i "s,/system,/system2,g" install-optional.sh
+	
+	cp -f ../alt-to-modify/%GAPPS% ../alt-to-modify/gapps_RDBSD_alt_%GAPPS%
+	zip -ruq ../alt-to-modify/gapps_RDBSD_alt_%GAPPS% *.*
+	mv -f ../alt-to-modify/gapps_RDBSD_alt_%GAPPS% ../alternate-Mod/gapps_RDBSD_alt_%GAPPS%
+	cd ..
+	)
+rm -fr tmp
 set PATH=%PREPATH%
 set PREPATH=
 echo.
@@ -194,19 +218,18 @@ goto :menu
 set PREPATH=%PATH%
 set PATH=%PATH%;.\tools;..\tools;..\..\tools
 cls
-echo .
-echo Clearing out recent mod
+echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.
+echo Clearing out recent mods
 sleep 3s
-rmdir /S /Q tmp\rd
-rmdir /S /Q tmp
-rmdir /S /Q pri-to-modify
-rmdir /S /Q alt-to-modify
-rmdir /S /Q Primary-Mod
-rmdir /S /Q Alternate-Mod
-mkdir pri-to-modify
-mkdir alt-to-modify
-mkdir Primary-Mod
-mkdir Alternate-Mod
+if exist tmp\nul rmdir /S /Q tmp
+if exist pri-to-modify\nul rmdir /S /Q pri-to-modify
+if exist alt-to-modify\nul rmdir /S /Q alt-to-modify
+if exist Primary-Mod\nul rmdir /S /Q Primary-Mod
+if exist Alternate-Mod\nul rmdir /S /Q Alternate-Mod
+if not exist pri-to-modify\nul mkdir pri-to-modify
+if not exist alt-to-modify\nul mkdir alt-to-modify
+if not exist Primary-Mod\nul mkdir Primary-Mod
+if not exist Alternate-Mod\nul mkdir Alternate-Mod
 set PATH=%PREPATH%
 set PREPATH=
 echo.
@@ -217,7 +240,8 @@ goto :menu
 :Exit
 set PATH=%PREPATH%
 set PREPATH=
-rmdir /S /Q tmp\rd
-rmdir /S /Q tmp
+if exist tmp\nul rmdir /S /Q tmp
+cls
+echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.&echo.
 echo Thank you for using Convert2Dualboot-SD!
-sleep 5s
+tools\sleep 5s
